requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.38
alembic>=1.14.1
pydantic>=2.10.0
pydantic-settings>=2.8.0
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
email-validator>=2.2.0
asyncpg>=0.30.0
boto3>=1.37.3
Pillow>=11.1.0
openai>=1.65.0
celery>=5.4.0
redis>=5.2.0
python-dotenv>=1.0.0
playwright>=1.50.0
beautifulsoup4>=4.13.0
nest_asyncio
